split out validator
